+++ packages/orq-rc/RELEASES.md
@@ -829,3 +829,13 @@ Based on:
 - [python v3.2.0-rc.45] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.45] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.45 - packages/orq-rc
+
+## 2025-03-13 10:36:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.517.3 (2.548.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.46] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.46] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.46 - packages/orq-rc